add a class to generate the go-mod and go-lic file
The go-mod code from the recipetool got moved to its own class
in openemmbedded-core named go-mod-update-modules.bbclass.

Unlike the go-mod-update-modules.bbclass in openembedded-core, this
uses `go install` instead of `go list` command to determine required
modules. It is a lot faster then the `go list` version, more then 8
mins vs 30 seconds or so for tailscale on my pc.

This seems to work for tailscale, as the comments indicate in the
go-mod-update-modules.bbclass the list command is used to avoid
pruning. go.mod files should nowadays also include indirect dependencies,
so this might be work in more general cases as well...

This adds an alternative task, do_update_modules_prune besides the
do_update_modules, which seems to work as well. License "detection"
remains a bit of an issue though.

NOTE: this is just for testing...